Fixed fastlane issues, added SimulatorStatusMagic for perfect status bars.
