+++ AutomatedScreenshotUITests/Info.plist
