+++ fastlane/README.md
@@ -20,7 +20,17 @@ or alternatively using `brew cask install fastlane`
 ```
 fastlane ios screenshots
 ```
-Generate new localized screenshots
+Generate app screenshots
+### ios frame
+```
+fastlane ios frame
+```
+Frame all screenshots
+### ios capture_and_frame
+```
+fastlane ios capture_and_frame
+```
+Capture and frame new screenshots
 
 ----
