+++ fastlane/Snapfile
@@ -5,9 +5,9 @@ devices([
    "iPhone 8",
 #   "iPhone 8 Plus",
 #   "iPhone SE",
-#   "iPhone XS",
 #   "iPhone XR",
-#   "iPhone XS Max",
+#   "iPhone XS",
+#   "iPhone XS Max"
 #   "iPad Pro (12.9-inch)",
 #   "iPad Pro (9.7-inch)",
 #   "Apple TV 1080p"
@@ -21,7 +21,7 @@ devices([
 # ])
 
 # The name of the scheme which contains the UI Tests
-scheme("FastlaneUIAutomation")
+scheme("AutomatedScreenshotUITests")
 
 # Where should the resulting screenshots be stored?
 output_directory("./fastlane/screenshots")
