+++ fastlane/report.xml
@@ -5,17 +5,17 @@
     
     
       
-      <testcase classname="fastlane.lanes" name="0: default_platform" time="0.007057">
+      <testcase classname="fastlane.lanes" name="0: update_fastlane" time="6.153408">
         
       </testcase>
     
       
-      <testcase classname="fastlane.lanes" name="1: capture_screenshots" time="104.889611">
+      <testcase classname="fastlane.lanes" name="1: default_platform" time="0.003047">
         
       </testcase>
     
       
-      <testcase classname="fastlane.lanes" name="2: frameit" time="0.053423">
+      <testcase classname="fastlane.lanes" name="2: capture_screenshots" time="53.257595">
         
       </testcase>
