+++ fastlane/screenshots/Framefile.json
@@ -13,6 +13,12 @@
     },
     "data": [
         {
+            "filter": "Home",
+            "keyword": {
+                "color": "#12274b"
+            }
+        },
+        {
             "filter": "Dining",
             "keyword": {
                 "color": "#12274b"
@@ -26,12 +32,6 @@
         },
         {
             "filter": "Laundry",
-            "keyword": {
-                "color": "#12274b"
-            }
-        },
-        {
-            "filter": "Fitness",
             "keyword": {
                 "color": "#12274b"
             }
