+++ fastlane/screenshots/en-US/keyword.strings
@@ -1,3 +1,4 @@
+"Home" = "HOME";
 "Dining" = "DINING";
 "GSR" = "GSR";
 "Laundry" = "LAUNDRY";
