+++ fastlane/screenshots/en-US/title.strings
@@ -1,5 +1,6 @@
+"Home" = "Everything you need."
 "Dining" = "Grab a bite.";
 "GSR" = "Book a room.";
 "Laundry" = "Find a machine.";
 "Fitness" = "Hit the gym.";
-"More" = "Penn news & links.";
+"More" = "News, fitness, and resources.";
