+++ fastlane/screenshots/screenshots.html
@@ -8,8 +8,26 @@
         font-family: "HelveticaNeue-Light", "Helvetica Neue Light", "Helvetica Neue", Helvetica, Arial, "Lucida Grande", sans-serif;
         font-weight: 300;
       }
-      .language {
-
+      #sortMenu {
+        overflow: hidden;
+        border: 1px solid #ccc;
+        background-color: #f1f1f1;
+        display: none;
+      }
+      #sortMenu button {
+        background-color: inherit;
+        float: left;
+        border: none;
+        outline: none;
+        cursor: pointer;
+        padding: 14px 16px;
+        font-size: 17px;
+      }
+      #sortMenu button:hover {
+        background-color: #ddd;
+      }
+      #sortMenu button.active {
+        background-color: #ccc;
       }
       .deviceName {
         display: block;
@@ -21,6 +39,14 @@
         cursor: pointer;
         border: 1px #EEE solid;
         z-index: 0;
+      }
+      .caption {
+        font-size: 24px;
+        padding-bottom: 24px;
+        padding-top: 30px;
+      }
+      h1, h2 {
+        font-weight: bold;
       }
       th {
         text-align: left;
@@ -65,6 +91,14 @@
     </style>
   </head>
   <body>
+    <div id="sortMenu">
+      <button id="defaultTab" class="tabLink" onclick="openTab(event, 'byLanguage')">By Language</button>
+      <button class="tabLink" onclick="openTab(event, 'byScreen')">By Screen</button>
+    </div>
+    <div id="byLanguage" class="tabContent"><h1 class="tabTitle">By Language:</h1>
+    </div>
+    <div id="byScreen" class="tabContent"><h1 class="tabTitle">By Screen:</h1>
+    </div>
     <div id="overlay">
       <img id="imageDisplay" src="" alt="" />
       <div id="imageInfo"></div>
@@ -74,6 +108,48 @@
       var imageDisplay   = document.getElementById('imageDisplay');
       var imageInfo      = document.getElementById('imageInfo');
       var screenshotLink = document.getElementsByClassName('screenshotLink');
+
+      window.onload = setup();
+
+      function setup() {
+        var i, menu, tabTitles;
+
+        // Since JS is enabled, show sort menu and hide tab titles
+        menu = document.getElementById("sortMenu");
+        menu.style.display = "block";
+
+        tabTitles = document.getElementsByClassName("tabTitle");
+        for (i = 0; i < tabTitles.length; i++) {
+          tabTitles[i].style.display = "none";
+        }
+
+        doClick(document.getElementById("defaultTab"));
+      }
+
+      function getCurrentTab() {
+        var i, tabs;
+        tabs = document.getElementsByClassName("tabContent");
+        for (i = 0; i < tabs.length; i++) {
+          if (tabs[i].style.display != "none") {
+            return i + 1;
+          }
+        }
+        return 1; // fallback
+      }
+
+      function openTab(evt, tabName) {
+        var i, tabContent, tabLinks;
+        tabs = document.getElementsByClassName("tabContent");
+        for (i = 0; i < tabs.length; i++) {
+          tabs[i].style.display = "none";
+        }
+        tabLinks = document.getElementsByClassName("tabLink");
+        for (i = 0; i < tabLinks.length; i++) {
+          tabLinks[i].className = tabLinks[i].className.replace(" active", "");
+        }
+        document.getElementById(tabName).style.display = "block";
+        evt.currentTarget.className += " active";
+      }
 
       function doClick(el) {
         if (document.createEvent) {
@@ -100,6 +176,7 @@
           tmpImg.src = img.src;
           imageDisplay.style.height     = 'auto';
           imageDisplay.style.width     = 'auto';
+          imageDisplay.style.paddingTop = 0;
           if (window.innerHeight < tmpImg.height) {
             imageDisplay.style.height = document.documentElement.clientHeight+'px';
           } else if (window.innerWidth < tmpImg.width) {
@@ -125,12 +202,13 @@
 
         overlay.style.display = "none";
 
+        img_tab = parseInt(getCurrentTab());
         img_counter = parseInt(e.target.dataset.counter) + 1;
         try {
-          link = document.body.querySelector('img[data-counter="'+img_counter+'"]').parentNode;
+          link = document.body.querySelector('img[data-tab="'+img_tab+'"][data-counter="'+img_counter+'"]').parentNode;
         } catch (e) {
           try {
-            link = document.body.querySelector('img[data-counter="0"]').parentNode;
+            link = document.body.querySelector('img[data-tab="'+img_tab+'"][data-counter="0"]').parentNode;
           } catch (e) {
             return false;
           }
